adjust wait time for EPS based on vehicle speed
it seems EPS will fault when the speed is low (e.g. <15mph) and LKAS torque is sent too soon. introduce a way to dyanmically adjust the wait time (cycles) based on the speed before spoofing

for speed <3mph, the wait time is 1s.
for speed <9mph, the wait time is 0.5s.
for speed <15mph, the wait time is 0.2s.
for speed <39mph, the wait time is 0.1s.
for speed >=39mph, no wait time.